app/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.2.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.3
dj-database-url==0.5.0
Django==3.1.2
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.5
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
iniconfig==1.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.1.1
pytest-cov==2.10.1
pytest-django==3.10.0
pytz==2020.1
regex==2020.9.27
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
whitenoise==5.2.0
